requirements.txt
djangorestframework
celery
flower